+++ http/mover.http
@@ -1,4 +1,4 @@
-@token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIyOTZhNmFjOC0yYjY3LTQ1ZTYtOWJlNi1kYTA1MzE5MjE1ZTUiLCJlbWFpbCI6InNreTEzQHRlc3QuY29tIiwibmFtZSI6IuydvOuwmOyduCIsInVzZXJUeXBlIjoiY2xpZW50IiwiaWF0IjoxNzUzNjMwMzQ5LCJleHAiOjE3NTM2MzM5NDl9.DmGYDU2QQQoMFtz1TuJbiEUrw_0OL1yyxGrHxBQccww
+@token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIyOTZhNmFjOC0yYjY3LTQ1ZTYtOWJlNi1kYTA1MzE5MjE1ZTUiLCJlbWFpbCI6InNreTEzQHRlc3QuY29tIiwibmFtZSI6IuydvOuwmOyduCIsInVzZXJUeXBlIjoiY2xpZW50IiwiaWF0IjoxNzUzNzA3OTkxLCJleHAiOjE3NTM3MTE1OTF9.EyR1M6-ZEg0EIUc_ge-TojOKslNxxIg5N9Hau-gkEvs
 @moverToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJiNjZkNTZkYS03YTE0LTRlMjgtYjRjMC1mZmNlZTdiYWJhMjkiLCJlbWFpbCI6Im1vdmVyMUBnbWFpbC5jb20iLCJuYW1lIjoi7LKg6rCA67CpIOyalOumrOyCrCIsInVzZXJUeXBlIjoibW92ZXIiLCJpYXQiOjE3NTM2MzExNzksImV4cCI6MTc1MzYzNDc3OX0.YGZ1zpqXgjGgvE2R0mc-7ijJ1USJp2iygigHhhKTUZ0
 @moverId=b66d56da-7a14-4e28-b4c0-ffcee7baba29
 ### 전체 기사님 리스트 조회 (비회원도 가능)
@@ -22,12 +22,11 @@ Authorization: Bearer {{moverToken}}
 Content-Type: application/json
 
 ### 기사님 지정 요청 (견적서와 매칭)
-PATCH http://localhost:4000/movers/d9da38fd-0b93-4fd9-9449-f4d29d1c864a
+PATCH http://localhost:4000/requests/movers/{{moverId}}
 Authorization: Bearer {{token}}
 Content-Type: application/json
 
 {
-  "requestId": "721a2ea4-56e2-4504-a0d6-5a7c0158e5c7",
-  "moverId": "d9da38fd-0b93-4fd9-9449-f4d29d1c864a"
+  "requestId": "296a6ac8-2b67-45e6-9be6-da05319215e5"
 }
