+++ src/routers/request.router.ts
@@ -13,4 +13,7 @@ requestRouter.get("/", verifyAccessToken, requestController.getReceivedRequests)
 // 받은 요청 조회 (일반 유저)
 requestRouter.get("/client/active", verifyAccessToken, requestController.getClientActiveRequests);
 
+// 기사님 지정 요청 (일반 > 기사)
+requestRouter.patch('/movers/:moverId', verifyAccessToken, requestController.designateMover);
+
 export default requestRouter;
